posts and simple_form to post added
