+++ app/controllers/communities_controller.rb
@@ -1,17 +1,8 @@
 class CommunitiesController < ApplicationController
-  def posts
-    @community = Community.find[params[:id]]
-    @post = Post.new
-  end
-
-  def create
-    @post = Post.new(post_params)
-    # we need `community_id` to associate post with corresponding community
-    @community = Community.find(params[:restaurant_id])
-    @post.community = @community
-    @post.save
-    redirect_to restaurant_path(@community)
-  end
+  # def posts
+  #   @community = Community.find[params[:id]]
+  #   @post = Post.new
+  # end
 
   def find_community
     @community = Community.find[params[:id]]
