+++ app/models/post.rb
@@ -1,4 +1,5 @@
 class Post < ApplicationRecord
   belongs_to :community
   has_many :comments
+  validates :content, presence: true
 end
